requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.0
google-generativeai==0.5.0
python-dotenv==1.0.1
reportlab==4.1.0

